use correct frame when importing a sav
